[FIX] 리파지토리 기본키 데이터 타입 변경

+++ src/main/java/com/example/parking/repository/SearchHistoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SearchHistoryRepository extends JpaRepository<SearchHistory, Integer> {
+public interface SearchHistoryRepository extends JpaRepository<SearchHistory, Long> {
     List<SearchHistory> findByUserOrderByHistoryIdDesc(User user);
 }
